+++ src/commands/project/__tests__/setArtifact.test.js
@@ -4,6 +4,14 @@
 
 import nock from 'nock'
 
+import {
+  notifiesWithUsageMessageForDashH,
+  notifiesWithUsageHintForInvalidArgs,
+  notifiesWithErrorIfNotAPlayer,
+  notifiesWithErrorForAPIErrors,
+  notifiesWithErrorForGraphQLErrors,
+} from '../../../../test/commonTests'
+
 describe(testContext(__filename), function () {
   describe('invoke', function () {
     before(function () {
@@ -15,7 +23,7 @@ describe(testContext(__filename), function () {
       }
       this.formatError = msg => `__FMT: ${msg}`
       this.lgJWT = 'not.a.real.token'
-      this.lgPlayer = {id: 'not.a.real.id'}
+      this.lgUser = {id: 'not.a.real.id', roles: ['player']}
     })
     beforeEach(function () {
       this.notifications = []
@@ -25,81 +33,24 @@ describe(testContext(__filename), function () {
       nock.cleanAll()
     })
 
-    it('notifies with the usage message when requested', function () {
-      const {lgJWT, lgPlayer} = this
-      return this.invoke(['-h'], this.notify, {lgJWT, lgPlayer})
-        .then(() => {
-          expect(this.notifications[0]).to.match(/Usage:/)
-        })
-    })
-
-    it('notifies with an error message when too few positional arguments are provided', function () {
-      const {lgJWT, lgPlayer} = this
-      return this.invoke(['#bad-lemurs-12'], this.notify, {lgJWT, lgPlayer})
-        .then(() => {
-          expect(this.notifications[0]).to.match(/wrong number of arguments/)
-        })
-    })
-
-    it('notifies with a warning if too many positional arguments are provided', function () {
-      const {lgJWT, lgPlayer} = this
-      return this.invoke(['#bad-lemurs-12', '#good-bears-2', 'http://example.com'], this.notify, {lgJWT, lgPlayer})
-        .then(() => {
-          expect(this.notifications[0]).to.match(/wrong number of arguments/)
-        })
-    })
-
+    it('notifies with the usage message when requested', notifiesWithUsageMessageForDashH)
+    it('notifies with an error message when too few positional arguments are provided', notifiesWithUsageHintForInvalidArgs(['#bad-lemurs-12']))
+    it('notifies with a warning if too many positional arguments are provided', notifiesWithUsageHintForInvalidArgs(['#bad-lemurs-12', '#good-bears-2', 'http://example.com']))
     it('notifies with an error message when invoked by a non-player', function () {
-      const {lgJWT} = this
-      return Promise.all([
-        this.invoke(['#bad-lemurs-12', 'http://example.com/'], this.notify, {lgJWT, lgPlayer: null})
-          .then(() => {
-            expect(this.notifications[0]).to.match(/not a player/)
-          }),
-        this.invoke(['#bad-lemurs-12', 'http://example.com/'], this.notify, {lgJWT, lgPlayer: {object: 'without id attribute'}})
-          .then(() => {
-            expect(this.notifications[1]).to.match(/not a player/)
-          })
-      ])
+      return notifiesWithErrorIfNotAPlayer(['#bad-lemurs-12', 'http://example.com/']).bind(this)()
     })
+    it('notifies of API invocation errors', notifiesWithErrorForAPIErrors(['#bad-lemurs-12', 'http://example.com/']))
+    it('notifies of GraphQL invocation errors', notifiesWithErrorForGraphQLErrors(['#bad-lemurs-12', 'http://example.com/']))
 
     it('notifies with a thank you message when the API invocation succeeds', function (done) {
       nock('http://game.learnersguild.test')
         .post('/graphql')
         .reply(200, {data: {setProjectArtifactURL: {id: '00000000-1111-2222-3333-444444444444'}}})
 
-      const {lgJWT, lgPlayer} = this
-      return this.invoke(['#bad-lemurs-12', 'http://example.com/'], this.notify, {lgJWT, lgPlayer})
+      const {lgJWT, lgUser} = this
+      return this.invoke(['#bad-lemurs-12', 'http://example.com/'], this.notify, {lgJWT, lgUser})
         .then(() => {
           expect(this.notifications[0]).to.match(/Thanks! The artifact for .+ is now set to .+/)
-          done()
-        })
-        .catch(err => done(err))
-    })
-
-    it('notifies of API invocation errors', function (done) {
-      nock('http://game.learnersguild.test')
-        .post('/graphql')
-        .reply(500, 'Internal Server Error')
-
-      const {lgJWT, lgPlayer, formatError} = this
-      return this.invoke(['#bad-lemurs-12', 'http://example.com/'], this.notify, {lgJWT, lgPlayer, formatError})
-        .then(() => {
-          expect(this.notifications[0]).to.equal('__FMT: Internal Server Error')
-          done()
-        })
-        .catch(err => done(err))
-    })
-
-    it('notifies of GraphQL invocation errors', function (done) {
-      nock('http://game.learnersguild.test')
-        .post('/graphql')
-        .reply(200, {errors: [{message: 'GraphQL Error'}]})
-
-      const {lgJWT, lgPlayer, formatError} = this
-      this.invoke(['#bad-lemurs-12', 'http://example.com/'], this.notify, {lgJWT, lgPlayer, formatError})
-        .then(() => {
-          expect(this.notifications[0]).to.equal('__FMT: GraphQL Error')
           done()
         })
         .catch(err => done(err))