+++ test/commonTests.js
@@ -13,8 +13,8 @@ export function notifiesWithUsageMessageForDashH() {
 
 export function notifiesWithUsageHintForInvalidArgs(invalidArgv) {
   return function () {  // don't use arrow-function b/c we need 'this'
-    const {lgJWT, lgPlayer} = this
-    return this.invoke(invalidArgv, this.notify, {lgJWT, lgPlayer})
+    const {lgJWT, lgUser} = this
+    return this.invoke(invalidArgv, this.notify, {lgJWT, lgUser})
       .then(() => {
         expect(this.notifications[0]).to.match(/\-\-help/)
       })
@@ -25,11 +25,11 @@ export function notifiesWithErrorIfNotAPlayer(validArgv) {
   return function () {  // don't use arrow-function b/c we need 'this'
     const {lgJWT} = this
     return Promise.all([
-      this.invoke(validArgv, this.notify, {lgJWT, lgPlayer: null})
+      this.invoke(validArgv, this.notify, {lgJWT, lgUser: null})
         .then(() => {
           expect(this.notifications[0]).to.match(/not a player/)
         }),
-      this.invoke(validArgv, this.notify, {lgJWT, lgPlayer: {object: 'without id attribute'}})
+      this.invoke(validArgv, this.notify, {lgJWT, lgUser: {id: 'not.a.real.id', roles: ['notplayer']}})
         .then(() => {
           expect(this.notifications[1]).to.match(/not a player/)
         })
@@ -45,7 +45,7 @@ export function notifiesWithErrorIfNotAModerator(validArgv) {
         .then(() => {
           expect(this.notifications[0]).to.match(/not a moderator/)
         }),
-      this.invoke(validArgv, this.notify, {lgJWT, lgUser: {roles: ['player']}})
+      this.invoke(validArgv, this.notify, {lgJWT, lgUser: {id: 'not.a.real.id', roles: ['player']}})
         .then(() => {
           expect(this.notifications[1]).to.match(/not a moderator/)
         })
@@ -59,9 +59,9 @@ export function notifiesWithErrorForAPIErrors(validArgv) {
       .post('/graphql')
       .reply(500, 'Internal Server Error')
 
-    const {lgJWT, lgPlayer, lgUser} = this
+    const {lgJWT, lgUser} = this
     const formatError = msg => `__FMT: ${msg}`
-    return this.invoke(validArgv, this.notify, {lgJWT, lgPlayer, lgUser, formatError})
+    return this.invoke(validArgv, this.notify, {lgJWT, lgUser, formatError})
       .then(() => {
         expect(this.notifications).to.include('__FMT: Internal Server Error')
         done()
@@ -76,9 +76,9 @@ export function notifiesWithErrorForGraphQLErrors(validArgv) {
       .post('/graphql')
       .reply(200, {errors: [{message: 'GraphQL Error'}]})
 
-    const {lgJWT, lgPlayer, lgUser} = this
+    const {lgJWT, lgUser} = this
     const formatError = msg => `__FMT: ${msg}`
-    this.invoke(validArgv, this.notify, {lgJWT, lgPlayer, lgUser, formatError})
+    this.invoke(validArgv, this.notify, {lgJWT, lgUser, formatError})
       .then(() => {
         expect(this.notifications).to.include('__FMT: GraphQL Error')
         done()